More updates to get the new name into the view model

+++ src/edu/jhuapl/saavtk/gui/ShapeModelImporter.java
@@ -3,22 +3,19 @@ package edu.jhuapl.saavtk.gui;
 import java.io.File;
 import java.io.IOException;
 import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Map.Entry;
 
 import org.apache.commons.io.FileUtils;
 
-import vtk.vtkPolyData;
-import vtk.vtkPolyDataWriter;
-import vtk.vtkSphereSource;
-import vtk.vtkTransform;
-import vtk.vtkTransformPolyDataFilter;
-import edu.jhuapl.saavtk.model.Graticule;
 import edu.jhuapl.saavtk.model.ShapeModel;
 import edu.jhuapl.saavtk.util.Configuration;
 import edu.jhuapl.saavtk.util.FileUtil;
 import edu.jhuapl.saavtk.util.MapUtil;
 import edu.jhuapl.saavtk.util.PolyDataUtil;
+import vtk.vtkPolyData;
+import vtk.vtkPolyDataWriter;
+import vtk.vtkSphereSource;
+import vtk.vtkTransform;
+import vtk.vtkTransformPolyDataFilter;
 
 public class ShapeModelImporter
 {
@@ -194,15 +191,18 @@ public class ShapeModelImporter
 
         System.out.println("ShapeModelImporter: importShapeModel: model path is " + modelPath);
         File jsonFile = new File(modelPath.substring(0, modelPath.length()-3) + "json");
-        File jsonFileDestination = new File(newModelDir.getAbsolutePath() + File.separator + "model.json");
-        
-        try {
-			FileUtil.copyFile(jsonFile, jsonFileDestination);
-		} catch (IOException e) {
-            errorMessage[0] = "The was an error copying " + jsonFile + " to " + jsonFileDestination;
-            e.printStackTrace();
-            return false;
-		}    
+        if (jsonFile.exists())
+        {
+	        File jsonFileDestination = new File(newModelDir.getAbsolutePath() + File.separator + "model.json");
+	        
+	        try {
+				FileUtil.copyFile(jsonFile, jsonFileDestination);
+			} catch (IOException e) {
+	            errorMessage[0] = "The was an error copying " + jsonFile + " to " + jsonFileDestination;
+	            e.printStackTrace();
+	            return false;
+			}    
+        }
 
         // Copy any files as needed
         for(String key : copyMap.keySet())
